+++ Q5/Rectangle.hpp
@@ -7,8 +7,8 @@ public:
 	string r_name="Rectangle";
 
 	Rectangle(const Point& center, const Size& size) :Shape(center,size){}
-
-	size_t area() const override {return size_.width *(size_t)size_.height;}
+    //change area function from size_t to double
+	double area() const override {return size_.width *(size_t)size_.height;}
 
 	string name() const override {return r_name;}
